[HW-11][Commit-10]-Java-Web. Fixes done

+++ src/main/java/com/academy/webacademy/FileRepository.java
@@ -3,24 +3,24 @@ package com.academy.webacademy;
 import java.io.*;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 class FileRepository {
 
-    private final File file = new File("/Users/efe/Desktop/web-academy/src/main/resources/db");
+    File pathToFiles = new File(Objects.requireNonNull(this.getClass().getClassLoader().getResource("db")).getPath());
 
     //TableServlet
     //GET
     public String listOfFiles() {
         List<String> arrayList = new ArrayList<>();
         System.out.println("I'm in listOfFiles in FileRedactor class and got here by GET /tables");
-        File[] files = file.listFiles();
+        File[] files = pathToFiles.listFiles();
         int counter = 0;
         assert files != null;
         for (File value : files) {
             if (value.isFile()) {
                 arrayList.add("File number " + ++counter + " : " + value.getName());
-            }
-            else if (value.isDirectory()) {
+            } else if (value.isDirectory()) {
                 return "Directory: " + value.getName();
             }
         }
@@ -30,14 +30,14 @@ class FileRepository {
     //TablesServlet
     //GET
     public String getFileContent(String path) throws IOException {
-        try (FileInputStream fileInputStream = new FileInputStream(file + path)) {
+        try (FileInputStream fileInputStream = new FileInputStream(pathToFiles + path)) {
             return new String(fileInputStream.readAllBytes());
         }
     }
 
     //POST
     public boolean createNewFile(String path, String body) throws IOException {
-        try(FileWriter fileWriter = new FileWriter(file + path)) {
+        try (FileWriter fileWriter = new FileWriter(pathToFiles + path)) {
             fileWriter.write(body);
             return true;
         }
@@ -45,15 +45,33 @@ class FileRepository {
 
     //PUT
     public boolean updateFileContent(String path, String name) throws IOException {
-        try (FileWriter fileWriter = new FileWriter(file + path, true)) {
-            fileWriter.write(name + " ");
-            return true;
+        File[] files = pathToFiles.listFiles();
+        assert files != null;
+        for (File file : files) {
+            if (file.getName().equals(path)) {
+                try (FileWriter fileWriter = new FileWriter(pathToFiles + path, true)) {
+                    fileWriter.write(name);
+                    return true;
+                }
+            } else {
+                throw new IOException();
+            }
         }
+        return true;
     }
 
     //DELETE
-    public boolean deleteFile(String path) {
-        File filePath = new File(file + path);
-        return filePath.delete();
+    public boolean deleteFile(String path) throws IOException {
+        File[] files = pathToFiles.listFiles();
+        assert files != null;
+        for (File file : files) {
+            if (file.getName().equals(path)) {
+                File filePath = new File(pathToFiles + path);
+                return filePath.delete();
+            } else {
+                throw new IOException();
+            }
+        }
+        return true;
     }
 }
